realizacion de validaciones en todas las entidades

+++ src/models/qualification.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, ForeignKey, DateTime,Float
 from sqlalchemy.orm import relationship
 from src.config.database import Base
 import datetime
@@ -7,7 +7,7 @@ class Qualification(Base):
     __tablename__ = "qualifications"
 
     id = Column(Integer, primary_key=True, index=True)
-    score = Column(Integer, nullable=False)
+    score = Column(Float, nullable=False)
     created_at= Column(DateTime, default=datetime.datetime.now)
     user_id = Column(Integer, ForeignKey('users.id'))
     parche_id = Column(Integer, ForeignKey('parches.id'))
